requirements.txt
absl-py==0.13.0
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astroid==2.11.2
astunparse==1.6.3
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
cffi==1.15.1
charset-normalizer==2.0.4
commonmark==0.9.1
cycler==0.10.0
debugpy==1.6.3
decorator==5.0.9
defusedxml==0.7.1
dill==0.3.4
entrypoints==0.4
fastjsonschema==2.16.2
flatbuffers==1.12
fvcore==0.1.5.post20220305
gast==0.4.0
google-auth==1.34.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
idna==3.2
imageio==2.9.0
importlib-metadata==4.6.3
importlib-resources==5.9.0
iopath==0.1.9
ipykernel==6.15.3
ipython==7.25.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
isort==5.10.1
jedi==0.18.0
Jinja2==3.1.2
joblib==1.0.1
jsonpatch==1.32
jsonpointer==2.2
jsonschema==4.16.0
jupyter==1.0.0
jupyter-client==7.3.5
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lazy-object-proxy==1.7.1
lxml==4.9.1
Markdown==3.3.4
MarkupSafe==2.1.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mccabe==0.7.0
mistune==2.0.4
nbclient==0.6.8
nbconvert==7.0.0
nbformat==5.5.0
nest-asyncio==1.5.5
networkx==2.6.3
notebook==6.4.12
numpy==1.21.5
oauthlib==3.1.1
onnx==1.11.0
onnx-simplifier==0.3.10
onnxoptimizer==0.2.7
onnxruntime==1.11.0
opencv-python==4.5.3.56
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.2
pandocfilters==1.5.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
pkgutil-resolve-name==1.3.10
platformdirs==2.5.1
portalocker==2.4.0
prometheus-client==0.14.1
prompt-toolkit==3.0.19
protobuf==3.17.3
psutil==5.9.2
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pycparser==2.21
Pygments==2.9.0
pylint==2.13.3
pyparsing==2.4.7
pyrsistent==0.18.1
python-dateutil==2.8.2
pytorch-model-summary==0.1.2
pytorch-wavelets @ file:///Users/mac/Documents/GitHub/wkd-datasets/pytorch_wavelets
pytz==2021.1
PyWavelets==1.1.1
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.3.2
QtPy==2.2.0
requests==2.26.0
requests-oauthlib==1.3.0
rich==12.4.4
rsa==4.7.2
scikit-image==0.18.3
scikit-learn==0.24.2
scipy==1.7.1
Send2Trash==1.8.0
setproctitle==1.3.2
six==1.15.0
sklearn==0.0
soupsieve==2.3.2.post1
tabulate==0.8.9
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4.1
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
terminado==0.15.0
threadpoolctl==2.2.0
tifffile==2021.10.12
tinycss2==1.1.1
tomli==2.0.1
torch==1.9.0
torch-tb-profiler==0.3.1
torchattacks==3.2.4
torchfile==0.1.0
torchvision==0.10.0
tornado==6.2
tqdm==4.62.3
traitlets==5.4.0
typed-ast==1.5.2
typing-extensions==4.1.1
urllib3==1.26.6
visdom==0.1.8.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
Werkzeug==2.0.1
widgetsnbextension==4.0.3
wrapt==1.12.1
yacs==0.1.8
zipp==3.5.0
